Trap errors when closing the tab
It is possible that change events can occur after the call to close the tab.
In that case, we get an invalid tab id error. Ignore that error because it
means we got the result we wanted and the tab is closed. Log any other type
of error that occurs. I have only seen invalid tab id, so I don't know what
other type of error is possible.